+++ mod-source-record-manager-server/src/test/java/org/folio/rest/impl/changeManager/ChangeManagerAPITest.java
@@ -89,6 +89,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
   private static final String CORRECT_RAW_RECORD_3 = "03401nam  22004091i 4500001001200000003000800012005001700020006001800037007001500055008004100070020003200111020003500143020004300178020004000221040002100261050002700282082001900309245016200328300002300490500004700513504005100560505178700611588004702398650003702445650004502482650002902527650003602556700005502592700006102647710002302708730001902731776005902750856011902809935001502928980003402943981001402977\u001Eybp10134220\u001ENhCcYBP\u001E20130220102526.4\u001Em||||||||d|||||||\u001Ecr||n|||||||||\u001E130220s2013    ncu     ob    001 0 eng d\u001E  \u001Fa1476601852 (electronic bk.)\u001E  \u001Fa9781476601854 (electronic bk.)\u001E  \u001Fz9780786471140 (softcover : alk. paper)\u001E  \u001Fz078647114X (softcover : alk. paper)\u001E  \u001FaNhCcYBP\u001FcNhCcYBP\u001E 4\u001FaPN1995.9.V46\u001FbG37 2013\u001E04\u001Fa791.43/656\u001F223\u001E00\u001FaGame on, Hollywood!\u001Fh[electronic resource] :\u001Fbessays on the intersection of video games and cinema /\u001Fcedited by Gretchen Papazian and Joseph Michael Sommers.\u001E  \u001Fa1 online resource.\u001E  \u001FaDescription based on print version record.\u001E  \u001FaIncludes bibliographical references and index.\u001E0 \u001FaIntroduction: manifest narrativity-video games, movies, and art and adaptation / Gretchen Papazian and Joseph Michael Sommers -- The rules of engagement: watching, playing and other narrative processes. Playing the Buffyverse, playing the gothic: genre, gender and cross-media interactivity in Buffy the vampire slayer: chaos bleeds / Katrin Althans -- Dead eye: the spectacle of torture porn in Dead rising / Deborah Mellamphy -- Playing (with) the western: classical Hollywood genres in modern video games / Jason W. Buel -- Game-to-film adaptation and how Prince of Persia: the sands of time negotiates the difference between player and audience / Ben S. Bunting, Jr -- Translation between forms of interactivity: how to build the better adaptation / Marcus Schulzke -- The terms of the tale: time, place and other ideologically constructed conditions. -- Playing (in) the city: the warriors and images of urban disorder / Aubrey Anable -- When did Dante become a scythe-wielding badass? modeling adaption and shifting gender convention in Dante's Inferno / Denise A. Ayo -- \"Some of this happened to the other fellow\": remaking Goldeneye with Daniel Craig / David McGowan -- Zombie stripper geishas in the new global economy: racism and sexism in video games / Stewart Chang -- Stories, stories everywhere (and nowhere just the same): transmedia texts. \"My name is Alan Wake. I'm a writer.\": crafting narrative complexity in the age of transmedia storytelling / Michael Fuchs -- Millions of voices: Star wars, digital games, fictional worlds and franchise canon / Felan Parker -- The hype man as racial stereotype, parody and ghost in Afro samurai / Treaandrea M. Russworm -- Epic nostalgia: narrative play and transmedia storytelling in Disney epic Mickey / Lisa K. Dusenberry.\u001E  \u001FaDescription based on print version record.\u001E 0\u001FaMotion pictures and video games.\u001E 0\u001FaFilm adaptations\u001FxHistory and criticism.\u001E 0\u001FaVideo games\u001FxAuthorship.\u001E 0\u001FaConvergence (Telecommunication)\u001E1 \u001FaPapazian, Gretchen,\u001Fd1968-\u001Feeditor of compilation.\u001E1 \u001FaSommers, Joseph Michael,\u001Fd1976-\u001Feeditor of  compilation.\u001E2 \u001FaEbooks Corporation\u001E0 \u001FaEbook Library.\u001E08\u001FcOriginal\u001Fz9780786471140\u001Fz078647114X\u001Fw(DLC)  2012051432\u001E40\u001FzConnect to e-book on Ebook Library\u001Fuhttp://qut.eblib.com.au.AU/EBLWeb/patron/?target=patron&extendedid=P_1126326_0\u001E  \u001Fa.o13465405\u001E  \u001Fa130307\u001Fb6000\u001Fe6000\u001Ff243967\u001Fg1\u001E  \u001FbOM\u001Fcnlnet\u001E\u001D";
   private static final String RAW_RECORD_RESULTING_IN_PARSING_ERROR = "01247nam  2200313zu 450000100110000000300080001100500170001905\u001F222\u001E1 \u001FaAriáes, Philippe.\u001E10\u001FaWestern attitudes toward death\u001Fh[electronic resource] :\u001Fbfrom the Middle Ages to the present /\u001Fcby Philippe Ariáes ; translated by Patricia M. Ranum.\u001E  \u001FaJohn Hopkins Paperbacks ed.\u001E  \u001FaBaltimore :\u001FbJohns Hopkins University Press,\u001Fc1975.\u001E  \u001Fa1 online resource.\u001E1 \u001FaThe Johns Hopkins symposia in comparative history ;\u001Fv4th\u001E  \u001FaDescription based on online resource; title from digital title page (viewed on Mar. 7, 2013).\u001E 0\u001FaDeath.\u001E2 \u001FaEbrary.\u001E 0\u001FaJohns Hopkins symposia in comparative history ;\u001Fv4th.\u001E40\u001FzConnect to e-book on Ebrary\u001Fuhttp://gateway.library.qut.edu.au/login?url=http://site.ebrary.com/lib/qut/docDetail.action?docID=10635130\u001E  \u001Fa.o1346565x\u001E  \u001Fa130307\u001Fb2095\u001Fe2095\u001Ff243966\u001Fg1\u001E  \u001FbOM\u001Fcnlnet\u001E\u001D\n";
   private static final String RAW_RECORD_RESULTING_IN_INSTANCE_MAPPING_ERROR = "04466cas a2200841 a 4500001000700000005001700007008004100024010004500065012002400110019004000134022003900174029002100213029001700234029002200251030002700273032001700300035002300317035008400340035001100424040038000435042002300815049000900838050001600847060001400863074002400877082001200901086001500913222005400928245005400982246000901036264012701045264007101172300003901243310002601282321004801308321002501356336002601381337002801407338002701435362004101462490004601503490003701549500011301586500005901699500005401758500006201812525004001874530002401914555003501938650007801973650002602051650002802077650002002105710008402125710008802209760003202297760003102329770010302360776011802463780007402581830008102655830006802736856007102804850024102875856025603116891003503372891004003407891004403447891003803491891004303529891004003572994001203612\u001E130824\u001E20171107114428.0\u001E741101c19409999ncusr p      f0   a0eng  \u001E  \u001Fa   40029172 \u001Fzsn 79004523 \u001Fzsn 82020472 \u001E  \u001Fb3\u001Fi8709\u001Fj3\u001Fk1\u001Flg\u001Fm1\u001E  \u001Fa1588389\u001Fa2264617\u001Fa4342024\u001Fa14349066\u001E0 \u001Fa0027-8874\u001Fl0027-8874\u001Fz0198-0157\u001F21\u001E1 \u001FaNLGGC\u001Fb842279245\u001E1 \u001FaNZ1\u001Fb4973626\u001E1 \u001FaAU@\u001Fb000023028302\u001E  \u001FaJNCIEQ\u001FzJNCIAM\u001FzJJIND8\u001E  \u001Fa001964\u001FbUSPS\u001E  \u001Fa(OCoLC)ocm01064763\u001E  \u001Fa(OCoLC)1064763\u001Fz(OCoLC)1588389\u001Fz(OCoLC)2264617\u001Fz(OCoLC)4342024\u001Fz(OCoLC)14349066\u001E  \u001Fa130824\u001E  \u001FaNLM\u001FcNLM\u001FdNSD\u001FdDLC\u001FdNSD\u001FdOCL\u001FdIUL\u001FdNSD\u001FdOCL\u001FdGPO\u001FdNSD\u001FdGPO\u001FdRCS\u001FdIUL\u001FdAIP\u001FdNST\u001FdAIP\u001FdDLC\u001FdOCL\u001FdNST\u001FdGPO\u001FdNSD\u001FdNST\u001FdHUL\u001FdNST\u001FdHUL\u001FdGPO\u001FdNSD\u001FdNLM\u001FdOCL\u001FdNST\u001FdOCL\u001FdNSD\u001FdGPO\u001FdNST\u001FdNSD\u001FdNST\u001FdWAU\u001FdNLM\u001FdNSD\u001FdGPA\u001FdNSD\u001FdNST\u001FdNSD\u001FdOCL\u001FdNST\u001FdGPO\u001FdDLC\u001FdOCL\u001FdGPO\u001FdMYG\u001FdNSD\u001FdGPO\u001FdGUA\u001FdGPO\u001FdMYG\u001FdHLS\u001FdGPO\u001FdOCL\u001FdDLC\u001FdGPO\u001FdOCLCQ\u001FdGPO\u001FdNSD\u001FdOCLCQ\u001FdDLC\u001FdGUA\u001FdOCL\u001FdMYG\u001FdNLGGC\u001FdEEM\u001FdOCLCQ\u001FdUtOrBLW\u001E  \u001Fansdp\u001Fapcc\u001Fapremarc\u001E  \u001FaALMM\u001E00\u001FaRC261\u001Fb.U47\u001E00\u001FaW1 JO941C\u001E  \u001Fa0488, 0488 (online)\u001E10\u001Fa616\u001F211\u001E0 \u001FaHE 20.3161\u001E 0\u001FaJournal of the National Cancer Institute\u001Fb(Print)\u001E00\u001FaJournal of the National Cancer Institute :\u001FbJNCI.\u001E30\u001FaJNCI\u001E 1\u001Fa[Bethesda, Md.] :\u001FbU.S. Department of Health, Education, and Welfare, Public Health Service, National Institutes of Health\u001E 2\u001Fa[Washington, D.C.] :\u001Fb[For sale by the Supt. of Docs., U.S. G.P.O]\u001E  \u001Favolumes :\u001Fbillustrations ;\u001Fc28 cm.\u001E  \u001FaSemimonthly,\u001Fb<1989->\u001E  \u001FaBimonthly (irregular),\u001FbAug. 1940-Apr. 1956\u001E  \u001FaMonthly,\u001FbJune 1956-\u001E  \u001Fatext\u001Fbtxt\u001F2rdacontent\u001E  \u001Faunmediated\u001Fbn\u001F2rdamedia\u001E  \u001Favolume\u001Fbnc\u001F2rdacarrier\u001E1 \u001FaBegan with: v. 1, no. 1 (Aug. 1940).\u001E1 \u001Fa-Apr. 1979: DHEW publication ;\u001Fvno. (NIH)\u001E1 \u001FaMay 1979-<1996>: NIH publication\u001E  \u001FaNo longer an official government publication; no longer distributed to depository libraries after Dec. 2002.\u001E  \u001FaPublished: Cary, NC : Oxford University Press, <2003->\u001E  \u001FaDescription based on: Vol. 62, no. 1 (Jan. 1979).\u001E  \u001FaLatest issue consulted: Vol. 95, no. 14 (July 16, 2003->.\u001E  \u001FaSupplements accompany some numbers.\u001E  \u001FaAlso issued online.\u001E  \u001FaVols. 1 (1940)-16 (1956). 1 v.\u001E 0\u001FaCancer\u001FvPeriodicals.\u001F0http://id.loc.gov/authorities/subjects/sh2008100056\u001E 2\u001FaNeoplasms\u001FvAbstracts.\u001E 2\u001FaNeoplasms\u001FvPeriodicals.\u001E17\u001FaOncologie.\u001F2gtt\u001E2 \u001FaNational Cancer Institute (U.S.)\u001F0http://id.loc.gov/authorities/names/n79107940\u001E2 \u001FaNational Institutes of Health (U.S.)\u001F0http://id.loc.gov/authorities/names/n78085445\u001E0 \u001FtDHEW publication\u001Fx0090-0206\u001E0 \u001FtNIH publication\u001Fx0276-4733\u001E0 \u001FtJournal of the National Cancer Institute. Monographs\u001Fx1052-6773\u001Fw(DLC)   94660018\u001Fw(OCoLC)21986096\u001E08\u001FiOnline version:\u001FtJournal of the National Cancer Institute (Online)\u001Fx1460-2105\u001Fw(DLC)  2005233013\u001Fw(OCoLC)42465676\u001E05\u001FtCancer treatment reports\u001Fx0361-5960\u001Fw(DLC)   76645250\u001Fw(OCoLC)2101497\u001E 0\u001FaDHEW publication ;\u001Fvno. (NIH)\u001F0http://id.loc.gov/authorities/names/n42009070\u001E 0\u001FaNIH publication.\u001F0http://id.loc.gov/authorities/names/n84710198\u001E41\u001Fuhttp://www.oxfordjournals.org/content?genre=journal&issn=0027-8874\u001E  \u001FaAU\u001FaCSt\u001FaCSt-L\u001FaCU-AM\u001FaCU-I\u001FaCU-RivA\u001FaCU-SM\u001FaCaOTU\u001FaCaOWtU\u001FaCaQMU\u001FaCoFS\u001FaCtY-M\u001FaDLC\u001FaFU-HC\u001FaGU\u001FaIaU\u001FaInU\u001FaMBCo\u001FaMBU-M\u001FaMCM\u001FaMH-BL\u001FaMH-CS\u001FaMMeT\u001FaMSobPR\u001FaNSyU\u001FaNcRS\u001FaOU\u001FaPPPCPh\u001FaPPiC\u001FaPPT\u001FaPU\u001FaPU-Med\u001FaRPB-S\u001FaTU-M\u001FaTxHMC\u001FaULA\u001FaViRCU-H\u001FaWaU\u001E41\u001Fahttp://purl.access.gpo.gov/GPO/LPS1716\u001Fzback issues are accessible by clicking on \"ABOUT THE JOURNAL\" and then clicking on \"ARCHIVE\" at the bottom of the screen (some issues available in: Full Text and Abstracts, Abstracts and PDF, and Abstracts Only)\u001E12\u001F9853\u001F81\u001Fav.\u001Fbno.\u001Fu06\u001Fvr\u001Fi(year)\u001E40\u001F9863\u001F81\u001Fa<10>-\u001Fi<1949>-\u001Fxprovisional\u001E12\u001F9853\u001F82\u001Fav.\u001Fbno.\u001Fu20\u001Fvr\u001Fi(year)\u001Fj(month)\u001E40\u001F9863\u001F82\u001Fa<80>\u001Fi<1988>\u001Fxprovisional\u001E12\u001F9853\u001F83\u001Fav.\u001Fbno.\u001Fu24\u001Fvr\u001Fi(year)\u001Fj(date)\u001E40\u001F9863\u001F83\u001Fa<81>-\u001Fi<1989>-\u001Fxprovisional\u001E  \u001FaC0\u001FbALM\u001E\u001D";
+  private static final String QUERY_PARAM_NAME = "defaultMapping";
 
   private Set<JobExecution.SubordinationType> parentTypes = EnumSet.of(
     JobExecution.SubordinationType.PARENT_SINGLE,
@@ -844,6 +845,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
   public void shouldReturnBadRequestOnPostWhenNoDtoPassedInBody() {
     RestAssured.given()
       .spec(spec)
+      .queryParam(QUERY_PARAM_NAME, true)
       .body(new JsonObject().toString())
       .when()
       .post(JOB_EXECUTION_PATH + UUID.randomUUID().toString() + RECORDS_PATH)
@@ -855,6 +857,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
   public void shouldReturnErrorOnPostRawRecordsWhenJobExecutionDoesNotExist() {
     RestAssured.given()
       .spec(spec)
+      .queryParam(QUERY_PARAM_NAME, true)
       .body(rawRecordsDto)
       .when()
       .post(JOB_EXECUTION_PATH + UUID.randomUUID().toString() + RECORDS_PATH)
@@ -891,6 +894,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
     async = testContext.async();
     RestAssured.given()
       .spec(spec)
+      .queryParam(QUERY_PARAM_NAME, true)
       .body(rawRecordsDto)
       .when()
       .post(JOB_EXECUTION_PATH + jobExec.getId() + RECORDS_PATH)
@@ -950,6 +954,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
     async = testContext.async();
     RestAssured.given()
       .spec(spec)
+      .queryParam(QUERY_PARAM_NAME, true)
       .body(rawRecordsDto_2)
       .when()
       .post(JOB_EXECUTION_PATH + jobExec.getId() + RECORDS_PATH)
@@ -1011,6 +1016,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
       async = testContext.async();
       RestAssured.given()
         .spec(spec)
+        .queryParam(QUERY_PARAM_NAME, true)
         .body(rawRecordsDto)
         .when()
         .post(JOB_EXECUTION_PATH + jobExec.getId() + RECORDS_PATH)
@@ -1083,6 +1089,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
     async = testContext.async();
     RestAssured.given()
       .spec(spec)
+      .queryParam(QUERY_PARAM_NAME, true)
       .body(rawRecordsDto)
       .when()
       .post(JOB_EXECUTION_PATH + jobExec.getId() + RECORDS_PATH)
@@ -1133,6 +1140,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
 
     RestAssured.given()
       .spec(spec)
+      .queryParam(QUERY_PARAM_NAME, true)
       .body(rawRecordsDto)
       .when()
       .post(JOB_EXECUTION_PATH + jobExec.getId() + RECORDS_PATH)
@@ -1170,6 +1178,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
     async = testContext.async();
     RestAssured.given()
       .spec(spec)
+      .queryParam(QUERY_PARAM_NAME, true)
       .body(rawRecordsDto)
       .when()
       .post(JOB_EXECUTION_PATH + jobExec.getId() + RECORDS_PATH)
@@ -1218,6 +1227,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
 
     RestAssured.given()
       .spec(spec)
+      .queryParam(QUERY_PARAM_NAME, true)
       .body(new RawRecordsDto()
         .withRecordsMetadata(new RecordsMetadata()
           .withContentType(RecordsMetadata.ContentType.MARC_RAW)
@@ -1274,6 +1284,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
     async = testContext.async();
     RestAssured.given()
       .spec(spec)
+      .queryParam(QUERY_PARAM_NAME, true)
       .body(rawRecordsDto)
       .when()
       .post(JOB_EXECUTION_PATH + jobExec.getId() + RECORDS_PATH)
@@ -1370,6 +1381,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
 
     RestAssured.given()
       .spec(spec)
+      .queryParam(QUERY_PARAM_NAME, true)
       .body(rawRecordsDto)
       .when()
       .post(JOB_EXECUTION_PATH + jobExec.getId() + RECORDS_PATH)
@@ -1425,6 +1437,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
     async = testContext.async();
     RestAssured.given()
       .spec(spec)
+      .queryParam(QUERY_PARAM_NAME, true)
       .body(rawRecordsDto)
       .when()
       .post(JOB_EXECUTION_PATH + jobExec.getId() + RECORDS_PATH)
@@ -1526,6 +1539,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
     async = testContext.async();
     RestAssured.given()
       .spec(spec)
+      .queryParam(QUERY_PARAM_NAME, true)
       .body(rawRecordsDtoContainingError)
       .when()
       .post(JOB_EXECUTION_PATH + jobExec.getId() + RECORDS_PATH)
@@ -1598,6 +1612,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
     async = testContext.async();
     RestAssured.given()
       .spec(spec)
+      .queryParam(QUERY_PARAM_NAME, true)
       .body(rawRecordsDtoContainingError)
       .when()
       .post(JOB_EXECUTION_PATH + jobExec.getId() + RECORDS_PATH)
@@ -1648,6 +1663,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
 
     RestAssured.given()
       .spec(spec)
+      .queryParam(QUERY_PARAM_NAME, true)
       .body(lastRawRecordsDto)
       .when()
       .post(JOB_EXECUTION_PATH + jobExec.getId() + RECORDS_PATH)
@@ -1717,6 +1733,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
     async = testContext.async();
     RestAssured.given()
       .spec(spec)
+      .queryParam(QUERY_PARAM_NAME, true)
       .body(rawRecordsDtoContainingError)
       .when()
       .post(JOB_EXECUTION_PATH + jobExec.getId() + RECORDS_PATH)
@@ -1798,6 +1815,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
     async = testContext.async();
     RestAssured.given()
       .spec(spec)
+      .queryParam(QUERY_PARAM_NAME,true)
       .body(rawRecordsDtoContainingError)
       .when()
       .post(JOB_EXECUTION_PATH + jobExec.getId() + RECORDS_PATH)
@@ -1850,6 +1868,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
     async = testContext.async();
     RestAssured.given()
       .spec(spec)
+      .queryParam(QUERY_PARAM_NAME, true)
       .body(rawRecordsDto)
       .when()
       .post(JOB_EXECUTION_PATH + jobExec.getId() + RECORDS_PATH)
@@ -1860,6 +1879,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
     async = testContext.async();
     RestAssured.given()
       .spec(spec)
+      .queryParam(QUERY_PARAM_NAME, true)
       .when()
       .delete(JOB_EXECUTION_PATH + jobExec.getId() + RECORDS_PATH)
       .then()
@@ -1881,6 +1901,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
     Async async = testContext.async();
     RestAssured.given()
       .spec(spec)
+      .queryParam(QUERY_PARAM_NAME, true)
       .when()
       .delete(JOB_EXECUTION_PATH + UUID.randomUUID() + RECORDS_PATH)
       .then()
@@ -1917,6 +1938,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
     async = testContext.async();
     RestAssured.given()
       .spec(spec)
+      .queryParam(QUERY_PARAM_NAME, true)
       .body(rawRecordsDto)
       .when()
       .post(JOB_EXECUTION_PATH + jobExec.getId() + RECORDS_PATH)
@@ -1927,6 +1949,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
     async = testContext.async();
     RestAssured.given()
       .spec(spec)
+      .queryParam(QUERY_PARAM_NAME, true)
       .when()
       .delete(JOB_EXECUTION_PATH + jobExec.getId() + RECORDS_PATH)
       .then()
@@ -1967,6 +1990,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
     async = testContext.async();
     RestAssured.given()
       .spec(spec)
+      .queryParam(QUERY_PARAM_NAME, true)
       .when()
       .delete(JOB_EXECUTION_PATH + jobExec.getId() + RECORDS_PATH)
       .then()
@@ -1983,4 +2007,53 @@ public class ChangeManagerAPITest extends AbstractRestTest {
     async.complete();
   }
 
+  @Test
+  public void shouldProcessChunkOfRawRecordsWhenQueryParamIsFalse(TestContext testContext) {
+    InitJobExecutionsRsDto response =
+      constructAndPostInitJobExecutionRqDto(1);
+    List<JobExecution> createdJobExecutions = response.getJobExecutions();
+    Assert.assertThat(createdJobExecutions.size(), is(1));
+    JobExecution jobExec = createdJobExecutions.get(0);
+
+    WireMock.stubFor(post(RECORDS_SERVICE_URL)
+      .willReturn(created().withTransformers(RequestToResponseTransformer.NAME)));
+
+    Async async = testContext.async();
+    RestAssured.given()
+      .spec(spec)
+      .body(new JobProfileInfo()
+        .withName("MARC records")
+        .withId(UUID.randomUUID().toString())
+        .withDataType(JobProfileInfo.DataType.MARC))
+      .when()
+      .put(JOB_EXECUTION_PATH + jobExec.getId() + JOB_PROFILE_PATH)
+      .then()
+      .statusCode(HttpStatus.SC_OK);
+    async.complete();
+
+    rawRecordsDto.getRecordsMetadata().setLast(true);
+    async = testContext.async();
+    RestAssured.given()
+      .spec(spec)
+      .queryParam(QUERY_PARAM_NAME, false)
+      .body(rawRecordsDto)
+      .when()
+      .post(JOB_EXECUTION_PATH + jobExec.getId() + RECORDS_PATH)
+      .then()
+      .statusCode(HttpStatus.SC_NO_CONTENT);
+    async.complete();
+
+    async = testContext.async();
+    RestAssured.given()
+      .spec(spec)
+      .when()
+      .get(JOB_EXECUTION_PATH + jobExec.getId())
+      .then()
+      .statusCode(HttpStatus.SC_OK)
+      .body("status", is(JobExecution.Status.PARSING_IN_PROGRESS.name()));
+    async.complete();
+
+    verify(1, postRequestedFor(urlEqualTo(PUBSUB_PUBLISH_URL)));
+    verify(1, postRequestedFor(urlEqualTo(RECORDS_SERVICE_URL)));
+  }
 }