+++ NEWS.md
@@ -7,7 +7,7 @@
 * Added mode of issuance mapping mechanism
 * Updated mapping for instance identifier types and unspecified instance type
 * Added JobExecutionProgress service
-* New endpoint for saving results of instance creation to journal was added 
+* Added defaultMapping query param to choose between default mapping and application of JobProfiles
 
  | METHOD |             URL                                    | DESCRIPTION                                         |
  |--------|----------------------------------------------------|-----------------------------------------------------|